Feat: notification for contact
