+++ package.json
@@ -22,6 +22,7 @@
     "react-markdown": "^9.0.3",
     "react-notion-x": "^7.2.6",
     "react-slick": "^0.30.3",
+    "react-toastify": "^11.0.3",
     "react-type-animation": "^3.2.0",
     "slick-carousel": "^1.8.1"
   },
