+++ src/sections/introduction/view/contact/Contact.css
@@ -1,0 +1,14 @@
+:root {
+  /* 성공 토스트 색상 */
+  --toastify-color-success: #00cccc;
+  /* 성공 토스트 텍스트 색상 */
+  --toastify-text-color-success: black;
+  --toastify-toast-width: 350px;
+  --toastify-toast-offset: 16px;
+}
+
+.Toastify__toast-container {
+  top: 80px;
+  right: 20px;
+  color: black;
+}
